CR-127: code cleanup and dummy data movement. (#619)
* deleted lr shared component moved dummydata to test folder


* Merge branch rack with EOS-29872


* lr bucket page completed


* lr style updates for sgtsvg component


* lr style updates for sgtsvgIcon component


* lr style fixes and code quality improvement


* deleted all shared component


* shortcut path added for lib


* removed the lib services
