+++ src/lib/components/SgtDropdown/SgtDropdown.model.ts
@@ -10,10 +10,11 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Affero General Public License for more details.
  * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ * along with this program. If not, see https://www.gnu.org/licenses/.
  * For any questions about this software or licensing,
  * please email opensource@seagate.com.
  */
+
 export interface SgtDropdownOption {
   label: string;
   value: any;
