+++ src/scss/style.scss
@@ -41,7 +41,6 @@ html {
   gap: 0.5em;
   margin-bottom: 0.5em;
 }
-.v-btn__content{
+.v-btn__content {
   text-transform: none;
-
 }
